chore: remove redundant testfile
