+++ contracts/test/foreign-proxy-arbitrum.test.js
@@ -10,16 +10,14 @@ const appealCost = 5000;
 const questionID = zeroPadValue(toBeHex(0), 32);
 const answer = zeroPadValue(toBeHex(11), 32);
 const arbitrationID = 0;
-const l2GasPrice = 100;
-const surplusAmount = 200; // Covers the gas price
-const totalCost = 1200; // Arbitration cost + surplus
+const surplusAmount = 20000; // Covers the gas price (gas limit * priceBid + ticketCost) and has some extra amount. Gas = 12500
+const totalCost = 21000; // Arbitration cost + surplus
 
-const L2_GAS_LIMIT = 1500000;
-const L2_GAS_PER_PUB_DATA_BYTE_LIMIT = 800;
-const l2BlockNumber = 15012;
-const messageIndex = 1;
-const l2TxNumberInBlock = 9;
-const proof = [zeroPadValue(toBeHex(11), 32)];
+const L2_GAS_LIMIT = 500;
+const gasPriceBid = 5;
+const l2GasPrice = 12500; // See above how this value calculated
+
+const ticketSubmissionCost = 10000;
 
 const appealTimeOut = 180;
 const winnerMultiplier = 3000;
@@ -39,7 +37,9 @@ let arbitrator;
 let homeProxy;
 let foreignProxy;
 let realitio;
-let mockZkSync;
+let mockInbox;
+let mockBridge;
+let mockOutbox;
 
 let governor;
 let requester;
@@ -51,7 +51,9 @@ let other;
 describe("Cross-chain arbitration with appeals", () => {
   beforeEach("initialize the contract", async function () {
     [governor, requester, crowdfunder1, crowdfunder2, answerer, other] = await ethers.getSigners();
-    ({ arbitrator, realitio, foreignProxy, homeProxy, mockZkSync } = await deployContracts(governor));
+    ({ arbitrator, realitio, foreignProxy, homeProxy, mockInbox, mockBridge, mockOutbox } = await deployContracts(
+      governor
+    ));
 
     // Create disputes so the index in tests will not be a default value.
     await arbitrator.connect(other).createDispute(42, arbitratorExtraData, { value: arbitrationCost });
@@ -63,79 +65,79 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should correctly set the initial values", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
-
     expect(await foreignProxy.arbitrator()).to.equal(arbitrator.target);
     expect(await foreignProxy.arbitratorExtraData()).to.equal(arbitratorExtraData);
-    expect(await foreignProxy.zkSyncAddress()).to.equal(mockZkSync.target);
-    expect(await foreignProxy.l2GasLimit()).to.equal(L2_GAS_LIMIT);
-    expect(await foreignProxy.l2GasPerPubdataByteLimit()).to.equal(L2_GAS_PER_PUB_DATA_BYTE_LIMIT);
-    expect(await foreignProxy.surplusAmount()).to.equal(200);
+    expect(await foreignProxy.inbox()).to.equal(mockInbox.target);
+    expect(await foreignProxy.l2GasLimit()).to.equal(500);
+    expect(await foreignProxy.gasPriceBid()).to.equal(5);
+    expect(await foreignProxy.surplusAmount()).to.equal(20000);
     expect(await foreignProxy.homeProxy()).to.equal(homeProxy.target);
-    expect(await foreignProxy.deployer()).to.equal(ZeroAddress);
     expect(await foreignProxy.wNative()).to.equal(other);
 
     expect(await homeProxy.metadata()).to.equal(metadata);
     expect(await homeProxy.foreignChainId()).to.equal(zeroPadValue(toBeHex(5), 32));
     expect(await homeProxy.foreignProxy()).to.equal(foreignProxy.target);
-    expect(await homeProxy.foreignProxyAlias()).to.equal(mockZkSync.target);
+    expect(await homeProxy.mockInbox()).to.equal(mockInbox.target);
+    expect(await homeProxy.mockBridge()).to.equal(mockBridge.target);
     expect(await homeProxy.realitio()).to.equal(realitio.target);
-
-    expect(await mockZkSync.l2TransactionBaseCost(0, 0, 0)).to.equal(l2GasPrice);
 
     // 0 - winner, 1 - loser, 2 - loserAppealPeriod.
     const multipliers = await foreignProxy.getMultipliers();
     expect(multipliers[0]).to.equal(3000);
     expect(multipliers[1]).to.equal(7000);
     expect(multipliers[2]).to.equal(5000);
-  });
 
-  it("Check setHomeProxy requires", async () => {
-    await expect(foreignProxy.connect(other).setHomeProxy(homeProxy.target)).to.be.revertedWith("Only deployer can");
+    expect(await mockOutbox.sender()).to.equal(homeProxy.target);
+    expect(await mockBridge.outbox()).to.equal(mockOutbox.target);
 
-    await foreignProxy.setHomeProxy(homeProxy.target);
-    await expect(foreignProxy.setHomeProxy(mockZkSync.target)).to.be.revertedWith("Only deployer can"); // Deployer is nullified so no one can re-set the proxy
-  });
-
-  it("Should not allow to request arbitration if home proxy is not set", async () => {
-    await expect(
-      foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost })
-    ).to.be.revertedWith("Home proxy is not set");
+    expect(await mockInbox.bridge()).to.equal(mockBridge.target);
+    expect(await mockInbox.submissionFee()).to.equal(10000);
   });
 
   it("Should set correct values when requesting arbitration and fire the event", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
-
     await expect(
       foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost })
     ).to.be.revertedWith("Deposit value too low");
 
+    const requesterAddress = await requester.getAddress();
     await expect(foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost }))
-      .to.emit(realitio, "MockNotifyOfArbitrationRequest")
-      .withArgs(questionID, await requester.getAddress())
-      .to.emit(homeProxy, "RequestNotified")
-      .withArgs(questionID, await requester.getAddress(), maxPrevious)
-      .to.emit(mockZkSync, "L2Request")
-      .withArgs(homeProxy.target, 0, L2_GAS_LIMIT, L2_GAS_PER_PUB_DATA_BYTE_LIMIT, await requester.getAddress())
+      .to.emit(mockInbox, "TicketSent")
+      .withArgs(
+        homeProxy.target,
+        0,
+        ticketSubmissionCost,
+        requesterAddress,
+        requesterAddress,
+        L2_GAS_LIMIT,
+        gasPriceBid
+      )
+      .to.emit(foreignProxy, "RetryableTicketCreated")
+      .withArgs(0) // TicketID
       .to.emit(foreignProxy, "ArbitrationRequested")
-      .withArgs(questionID, await requester.getAddress(), maxPrevious);
+      .withArgs(questionID, requesterAddress, maxPrevious);
 
     const arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
     expect(arbitration[0]).to.equal(1, "Incorrect status of the arbitration after creating a request");
-    expect(arbitration[1]).to.equal(1100, "Deposit value stored incorrectly");
+    expect(arbitration[1]).to.equal(8500, "Deposit value stored incorrectly"); // Surplus (20000) + ArbCost (1000) - ArbitrumFee (10000 + 5*500).
 
-    const request = await homeProxy.requests(questionID, await requester.getAddress());
+    // Inbox acts as foreignProxyAlias here so it passes onlyForeignProxyAlias require
+    await expect(mockInbox.connect(other).redeemTicket(0))
+      .to.emit(realitio, "MockNotifyOfArbitrationRequest")
+      .withArgs(questionID, requesterAddress)
+      .to.emit(homeProxy, "RequestNotified")
+      .withArgs(questionID, requesterAddress, maxPrevious);
+
+    const request = await homeProxy.requests(questionID, requesterAddress);
     expect(request[0]).to.equal(2, "Incorrect status of the request in HomeProxy");
     expect(request[1]).to.equal(ZERO_HASH, "Answer should be empty");
 
     expect(await homeProxy.questionIDToRequester(questionID)).to.equal(
-      await requester.getAddress(),
+      requesterAddress,
       "Incorrect requester stored in home proxy"
     );
   });
 
   it("Should not allow to request arbitration 2nd time", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
 
     await expect(
@@ -143,8 +145,7 @@ describe("Cross-chain arbitration with appeals", () => {
     ).to.be.revertedWith("Arbitration already requested");
   });
 
-  it("Should have correct balance after paying arbitration cost and zk gas fee", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
+  it("Should have correct balance after paying arbitration cost and arbitrum fee", async () => {
     const oldBalance = await getBalance(requester);
 
     const tx = await foreignProxy
@@ -154,7 +155,7 @@ describe("Cross-chain arbitration with appeals", () => {
 
     const newBalance = await getBalance(requester);
     expect(newBalance).to.equal(
-      oldBalance - toBigInt(1000) - txFee - toBigInt(200), // Subtract tx fee, arbitration cost and surplus. The leftover surplus will be reimbursed later
+      oldBalance - toBigInt(1000) - txFee - toBigInt(20000), // Subtract tx fee, arbitration cost and surplus. The leftover surplus will be reimbursed later
       "Requester was not reimbursed correctly"
     );
   });
@@ -176,39 +177,25 @@ describe("Cross-chain arbitration with appeals", () => {
   it("Check foreign proxy permissions", async () => {
     await expect(
       foreignProxy.receiveArbitrationAcknowledgement(questionID, await requester.getAddress())
-    ).to.be.revertedWith("Only L2 allowed");
+    ).to.be.revertedWith("NOT_BRIDGE");
 
     await expect(
       foreignProxy.receiveArbitrationCancelation(questionID, await requester.getAddress())
-    ).to.be.revertedWith("Only L2 allowed");
+    ).to.be.revertedWith("NOT_BRIDGE");
   });
 
   it("Should set correct values when acknowledging arbitration and create a dispute", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-
-    const handleRequestPromise = homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestTx = await handleRequestPromise;
-    const handleRequestReceipt = await handleRequestTx.wait();
-
-    await expect(handleRequestPromise)
-      .to.emit(homeProxy, "RequestAcknowledged")
-      .withArgs(questionID, await requester.getAddress());
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    const request = await homeProxy.requests(questionID, await requester.getAddress());
-    expect(request[0]).to.equal(3, "Incorrect status of the request in HomeProxy");
+    await mockInbox.connect(other).redeemTicket(0);
 
     const badMessage = "0xfa";
-    await expect(
-      foreignProxy
-        .connect(other)
-        .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, badMessage, proof)
-    ).to.be.revertedWith("Could not receive L2 call");
+    await expect(mockBridge.connect(other).sendAsBridge(homeProxy.target, badMessage)).to.be.revertedWith(
+      "Failed TxToL1"
+    );
 
-    await expect(
-      foreignProxy.connect(other).consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof)
-    )
+    await expect(homeProxy.connect(other).handleNotifiedRequest(questionID, await requester.getAddress()))
+      .to.emit(homeProxy, "RequestAcknowledged")
+      .withArgs(questionID, await requester.getAddress())
       .to.emit(arbitrator, "DisputeCreation")
       .withArgs(2, foreignProxy.target)
       .to.emit(foreignProxy, "ArbitrationCreated")
@@ -216,10 +203,8 @@ describe("Cross-chain arbitration with appeals", () => {
       .to.emit(foreignProxy, "Dispute")
       .withArgs(arbitrator.target, 2, 0, 0);
 
-    expect(await foreignProxy.isL2ToL1MessageProcessed(l2BlockNumber, messageIndex)).to.equal(
-      true,
-      "Message should be marked as processed"
-    );
+    const request = await homeProxy.requests(questionID, await requester.getAddress());
+    expect(request[0]).to.equal(3, "Incorrect status of the request in HomeProxy");
 
     const arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
     expect(arbitration[0]).to.equal(2, "Incorrect status of the arbitration after acknowledging arbitration");
@@ -251,30 +236,31 @@ describe("Cross-chain arbitration with appeals", () => {
     expect(dispute[2]).to.equal(1000, "Incorrect fees value stored");
   });
 
-  it("Should not be able to proccess the message twice", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
+  it("Should not allow to receive the message from incorrect L2 sender", async () => {
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await mockInbox.connect(other).redeemTicket(0);
+    // Deliberately change outbox sender to see if the check in onlyL2Bridge modifier works.
+    // Note that it want obtain the exact message during revert because of low level call
+    await mockOutbox.setSender(await other.getAddress());
 
     await expect(
-      foreignProxy.consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof)
-    ).to.be.revertedWith("Message already processed");
+      homeProxy.connect(other).handleNotifiedRequest(questionID, await requester.getAddress())
+    ).to.be.revertedWith("Failed TxToL1");
+  });
+
+  it("Should not be able to proccess the message twice", async () => {
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
+    await mockInbox.connect(other).redeemTicket(0);
+
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await expect(
       foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost })
     ).to.be.revertedWith("Dispute already created");
   });
 
-  it("Should not allow to handle request before foreign proxy", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
+  it("Should not allow to handle request before ticket is redeemed", async () => {
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
 
     await expect(homeProxy.handleNotifiedRequest(questionID, await requester.getAddress())).to.be.revertedWith(
       "Invalid request status"
@@ -282,22 +268,15 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should reimburse after dispute creation in case of overpay", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: oneETH }); // Deliberately overpay
+    await mockInbox.connect(other).redeemTicket(0);
 
     const arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
     const requesterDeposit = arbitration[1];
-    expect(requesterDeposit).to.equal(oneETH - toBigInt(l2GasPrice), "Incorrect deposit value");
-
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
+    expect(requesterDeposit).to.equal(oneETH - toBigInt(12500), "Incorrect deposit value"); //Ticket submission fee + gas limit * price bid = 10000 * 2500
 
     const oldBalance = await getBalance(requester);
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     const newBalance = await getBalance(requester);
     expect(newBalance).to.equal(
@@ -306,37 +285,17 @@ describe("Cross-chain arbitration with appeals", () => {
     );
   });
 
-  it("Requester should be reimbursed leftover surplus after dispute creation if arbitration cost did not change", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    const oldBalance = await getBalance(requester);
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
-
-    const newBalance = await getBalance(requester);
-    expect(newBalance).to.equal(oldBalance + toBigInt(surplusAmount - l2GasPrice), "Balance should stay the same");
-  });
-
   it("Should cancel arbitration correctly", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
-
     await expect(homeProxy.handleRejectedRequest(questionID, await requester.getAddress())).to.be.revertedWith(
       "Invalid request status"
     );
 
     const badMaxPrevious = 11;
-    await expect(foreignProxy.connect(requester).requestArbitration(questionID, badMaxPrevious, { value: totalCost }))
+    await foreignProxy.connect(requester).requestArbitration(questionID, badMaxPrevious, { value: totalCost });
+
+    await expect(mockInbox.connect(other).redeemTicket(0))
       .to.emit(homeProxy, "RequestRejected")
-      .withArgs(questionID, await requester.getAddress(), badMaxPrevious, "Bond has changed")
-      .to.emit(foreignProxy, "ArbitrationRequested")
-      .withArgs(questionID, await requester.getAddress(), badMaxPrevious);
+      .withArgs(questionID, await requester.getAddress(), badMaxPrevious, "Bond has changed");
 
     let request = await homeProxy.requests(questionID, await requester.getAddress());
     expect(request[0]).to.equal(1, "Incorrect status of the request in HomeProxy after rejection");
@@ -345,31 +304,21 @@ describe("Cross-chain arbitration with appeals", () => {
       "Requester address should be empty after rejection"
     );
 
-    const handleRejectedRequestPromise = homeProxy.handleRejectedRequest(questionID, await requester.getAddress());
-    const handleRejectedRequestTx = await handleRejectedRequestPromise;
-    const handleRejectedRequestReceipt = await handleRejectedRequestTx.wait();
-
-    await expect(handleRejectedRequestPromise)
+    const oldBalance = await getBalance(requester);
+    await expect(homeProxy.handleRejectedRequest(questionID, await requester.getAddress()))
       .to.emit(homeProxy, "RequestCanceled")
+      .withArgs(questionID, await requester.getAddress())
+      .to.emit(foreignProxy, "ArbitrationCanceled")
       .withArgs(questionID, await requester.getAddress());
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRejectedRequestReceipt, homeProxy.interface).args;
 
     request = await homeProxy.requests(questionID, await requester.getAddress());
     expect(request[0]).to.equal(0, "Status should be nullified in home proxy");
 
-    const oldBalance = await getBalance(requester);
-    await expect(
-      foreignProxy.connect(other).consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof)
-    )
-      .to.emit(foreignProxy, "ArbitrationCanceled")
-      .withArgs(questionID, await requester.getAddress());
-
     const newBalance = await getBalance(requester);
     expect(newBalance).to.equal(
-      oldBalance + toBigInt(arbitrationCost + surplusAmount - l2GasPrice),
+      oldBalance + toBigInt(arbitrationCost + surplusAmount - 12500), // 12500 is Arbitrum fee
       "Requester was not reimbursed correctly"
-    ); // 1100
+    );
 
     const arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
     expect(arbitration[0]).to.equal(0, "Status should be empty");
@@ -378,29 +327,18 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should correctly handle failed dispute creation", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
+    await mockInbox.connect(other).redeemTicket(0);
 
     await expect(
       foreignProxy
         .connect(other)
-        .handleFailedDisputeCreation(questionID, await requester.getAddress(), { value: l2GasPrice })
+        .handleFailedDisputeCreation(questionID, await requester.getAddress(), { value: totalCost })
     ).to.be.revertedWith("Invalid arbitration status");
 
-    await arbitrator.setArbitrationPrice(2000); // Increase the cost so creation fails.
+    await arbitrator.setArbitrationPrice(8501); // Increase the cost so creation fails. 21000 was originally sent, 12500 taken as fee, so increased cost should be higher than 8500
 
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    await expect(handleRequestPromise)
-      .to.emit(homeProxy, "RequestAcknowledged")
-      .withArgs(questionID, await requester.getAddress());
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await expect(
-      foreignProxy.connect(other).consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof)
-    )
+    await expect(homeProxy.handleNotifiedRequest(questionID, await requester.getAddress()))
       .to.emit(foreignProxy, "ArbitrationFailed")
       .withArgs(questionID, await requester.getAddress());
 
@@ -414,20 +352,23 @@ describe("Cross-chain arbitration with appeals", () => {
         .connect(other)
         .handleFailedDisputeCreation(questionID, await requester.getAddress(), { value: l2GasPrice })
     )
+      .to.emit(foreignProxy, "RetryableTicketCreated")
+      .withArgs(1)
+      .to.emit(foreignProxy, "ArbitrationCanceled")
+      .withArgs(questionID, await requester.getAddress())
+      .to.emit(mockInbox, "TicketSent");
+
+    await expect(mockInbox.connect(other).redeemTicket(1))
       .to.emit(realitio, "MockCancelArbitrationRequest")
       .withArgs(questionID)
       .to.emit(homeProxy, "ArbitrationFailed")
-      .withArgs(questionID, await requester.getAddress())
-      .to.emit(mockZkSync, "L2Request")
-      .withArgs(homeProxy.target, 0, L2_GAS_LIMIT, L2_GAS_PER_PUB_DATA_BYTE_LIMIT, await other.getAddress())
-      .to.emit(foreignProxy, "ArbitrationCanceled")
       .withArgs(questionID, await requester.getAddress());
 
     const newBalance = await getBalance(requester);
     expect(newBalance).to.equal(
       oldBalance + toBigInt(arbitrationCost + surplusAmount - l2GasPrice),
       "Requester was not reimbursed correctly"
-    ); // 1100
+    ); // 8500
 
     arbitration = await foreignProxy.arbitrationRequests(0, await requester.getAddress());
     expect(arbitration[0]).to.equal(5, "Status should be Failed");
@@ -439,25 +380,16 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should correctly reimburse the overpay when handling failed dispute", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-
-    await arbitrator.setArbitrationPrice(2000); // Increase the cost so creation fails.
-
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await mockInbox.connect(other).redeemTicket(0);
+    await arbitrator.setArbitrationPrice(8501); // Increase the cost so creation fails.
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await expect(
       foreignProxy
         .connect(other)
         .handleFailedDisputeCreation(questionID, await requester.getAddress(), { value: l2GasPrice - 1 })
-    ).to.be.revertedWith("Should cover the zk fee");
+    ).to.be.revertedWith("Should cover arbitrum fee");
 
     const oldBalance = await getBalance(other);
     const tx = await foreignProxy
@@ -465,7 +397,7 @@ describe("Cross-chain arbitration with appeals", () => {
       .handleFailedDisputeCreation(questionID, await requester.getAddress(), { gasPrice: gasPrice, value: oneETH });
     const txFee = (await tx.wait()).gasUsed * gasPrice;
 
-    const newBalance = await await getBalance(other);
+    const newBalance = await getBalance(other);
     expect(newBalance).to.equal(
       oldBalance - toBigInt(l2GasPrice) - txFee, // Take only gas price for L2 and txfee
       "Caller was not reimbursed correctly"
@@ -473,18 +405,10 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should handle the ruling correctly", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
+    await mockInbox.connect(other).redeemTicket(0);
 
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
-    await expect(foreignProxy.rule(2, 8)).to.be.revertedWith("Only arbitrator allowed");
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     const arbAnswer = zeroPadValue(toBeHex(7), 32);
 
@@ -503,9 +427,14 @@ describe("Cross-chain arbitration with appeals", () => {
     );
 
     await expect(foreignProxy.connect(other).relayRule(questionID, await requester.getAddress(), { value: l2GasPrice }))
-      .to.emit(homeProxy, "ArbitratorAnswered")
-      .withArgs(questionID, arbAnswer)
+      .to.emit(foreignProxy, "RetryableTicketCreated")
+      .withArgs(1)
+      .to.emit(mockInbox, "TicketSent")
       .to.emit(foreignProxy, "RulingRelayed")
+      .withArgs(questionID, arbAnswer);
+
+    await expect(mockInbox.connect(other).redeemTicket(1))
+      .to.emit(homeProxy, "ArbitratorAnswered")
       .withArgs(questionID, arbAnswer);
 
     arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
@@ -526,22 +455,15 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should correctly reimburse the overpay when ruling is relayed", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
+    await mockInbox.connect(other).redeemTicket(0);
 
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
     await arbitrator.giveRuling(2, 8);
 
     await expect(
       foreignProxy.connect(other).relayRule(questionID, await requester.getAddress(), { value: l2GasPrice - 1 })
-    ).to.be.revertedWith("Should cover the zk fee");
+    ).to.be.revertedWith("Should cover arbitrum fee");
 
     const oldBalance = await getBalance(other);
     const tx = await foreignProxy
@@ -564,21 +486,14 @@ describe("Cross-chain arbitration with appeals", () => {
     let tx;
     let txFee;
     let roundInfo;
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
+    await mockInbox.connect(other).redeemTicket(0);
 
     await expect(foreignProxy.connect(crowdfunder1).fundAppeal(arbitrationID, 11, { value: 1000 })).to.be.revertedWith(
       "No dispute to appeal."
     );
 
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     // Check that can't fund the dispute that is not appealable.
     await expect(foreignProxy.connect(crowdfunder1).fundAppeal(arbitrationID, 11, { value: 1000 })).to.be.revertedWith(
@@ -658,16 +573,9 @@ describe("Cross-chain arbitration with appeals", () => {
 
   it("Should correctly create and fund subsequent appeal rounds", async () => {
     let roundInfo;
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await mockInbox.connect(other).redeemTicket(0);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 21, appealCost, appealTimeOut);
 
@@ -709,16 +617,9 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should not fund the appeal after the timeout", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await mockInbox.connect(other).redeemTicket(0);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 21, appealCost, appealTimeOut);
 
@@ -748,16 +649,9 @@ describe("Cross-chain arbitration with appeals", () => {
     const crowdfunder1Address = await crowdfunder1.getAddress();
     const crowdfunder2Address = await crowdfunder2.getAddress();
 
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await mockInbox.connect(other).redeemTicket(0);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 5, appealCost, appealTimeOut);
 
@@ -898,7 +792,10 @@ describe("Cross-chain arbitration with appeals", () => {
 
     newBalance = await getBalance(requester);
     newBalance2 = await getBalance(crowdfunder2);
-    expect(newBalance).to.equal(oldBalance + toBigInt(8499), "The balance of the requester is incorrect (withdraw 2 round)");
+    expect(newBalance).to.equal(
+      oldBalance + toBigInt(8499),
+      "The balance of the requester is incorrect (withdraw 2 round)"
+    );
     expect(newBalance2).to.equal(
       oldBalance2 + toBigInt(6500),
       "The balance of the crowdfunder2 is incorrect (withdraw 2 round)"
@@ -912,16 +809,9 @@ describe("Cross-chain arbitration with appeals", () => {
     let oldBalance;
     let newBalance;
 
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await mockInbox.connect(other).redeemTicket(0);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 20, appealCost, appealTimeOut);
 
@@ -963,16 +853,9 @@ describe("Cross-chain arbitration with appeals", () => {
     let oldBalance;
     let newBalance;
 
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await mockInbox.connect(other).redeemTicket(0);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 3, appealCost, appealTimeOut);
 
@@ -1014,16 +897,9 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should switch the ruling if the loser paid appeal fees while winner did not", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
-
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
+    await mockInbox.connect(other).redeemTicket(0);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 14, appealCost, appealTimeOut);
 
@@ -1037,36 +913,41 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should correctly submit evidence", async () => {
-    await foreignProxy.setHomeProxy(homeProxy.target);
     await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    const handleRequestPromise = await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    const handleRequestReceipt = await handleRequestPromise.wait();
-    const [l2Message] = getEmittedEvent("L1MessageSent", handleRequestReceipt, homeProxy.interface).args;
+    await mockInbox.connect(other).redeemTicket(0);
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
-    await foreignProxy
-      .connect(other)
-      .consumeMessageFromL2(l2BlockNumber, messageIndex, l2TxNumberInBlock, l2Message, proof);
     await expect(foreignProxy.connect(other).submitEvidence(arbitrationID, "text"))
       .to.emit(foreignProxy, "Evidence")
       .withArgs(arbitrator.target, arbitrationID, await other.getAddress(), "text");
   });
 
-  function getEmittedEvent(eventName, receipt, iface) {
-    return receipt.logs.map(log => iface.parseLog(log)).find(parsed => parsed.name === eventName);
-  }
-
   async function deployContracts(signer) {
     const Arbitrator = await ethers.getContractFactory("AutoAppealableArbitrator", signer);
     const arbitrator = await Arbitrator.deploy(String(arbitrationCost));
 
-    const MockZkSync = await ethers.getContractFactory("MockZkSync", signer);
-    const mockZkSync = await MockZkSync.deploy(l2GasPrice);
+    const MockOutbox = await ethers.getContractFactory("MockOutbox", signer);
+    const mockOutbox = await MockOutbox.deploy();
+
+    const MockBridge = await ethers.getContractFactory("MockBridge", signer);
+    const mockBridge = await MockBridge.deploy(mockOutbox.target);
+
+    const MockInbox = await ethers.getContractFactory("MockInbox", signer);
+    const mockInbox = await MockInbox.deploy(mockBridge.target, ticketSubmissionCost);
 
     const Realitio = await ethers.getContractFactory("MockRealitio", signer);
     const realitio = await Realitio.deploy();
 
-    const ForeignProxy = await ethers.getContractFactory("RealitioForeignProxyZkSync", signer);
-    const HomeProxy = await ethers.getContractFactory("MockRealitioHomeProxyZkSync", signer);
+    const ForeignProxy = await ethers.getContractFactory("RealitioForeignProxyArbitrum", signer);
+    const HomeProxy = await ethers.getContractFactory("MockRealitioHomeProxyArbitrum", signer);
+
+    const address = await signer.getAddress();
+    const nonce = await signer.getNonce();
+
+    const homeProxyAddress = ethers.getCreateAddress({
+      from: address,
+      nonce: nonce + 1, // Add 1 since homeProxy deployment will be after foreignProxy
+    });
 
     const foreignProxy = await ForeignProxy.deploy(
       other, // Use other address as placeholder for wNative
@@ -1076,26 +957,31 @@ describe("Cross-chain arbitration with appeals", () => {
       winnerMultiplier,
       loserMultiplier,
       loserAppealPeriodMultiplier,
-      mockZkSync.target,
-      L2_GAS_LIMIT,
-      L2_GAS_PER_PUB_DATA_BYTE_LIMIT,
+      homeProxyAddress,
+      mockInbox.target,
       surplusAmount,
+      [L2_GAS_LIMIT, gasPriceBid]
     );
 
     const homeProxy = await HomeProxy.deploy(
       realitio.target,
       metadata,
       foreignProxy.target,
-      mockZkSync.target, // Actual alias would be derivative of foreignProxy address and is handled by zkSync
-      foreignChainId
+      foreignChainId,
+      mockInbox.target,
+      mockBridge.target
     );
+
+    await mockOutbox.setSender(homeProxy.target);
 
     return {
       arbitrator,
       realitio,
       foreignProxy,
       homeProxy,
-      mockZkSync,
+      mockInbox,
+      mockBridge,
+      mockOutbox,
     };
   }
