+++ contracts/test/foreign-proxy-polygon.test.js
@@ -10,14 +10,6 @@ const appealCost = 5000;
 const questionID = zeroPadValue(toBeHex(0), 32);
 const answer = zeroPadValue(toBeHex(11), 32);
 const arbitrationID = 0;
-const surplusAmount = 20000; // Covers the gas price (gas limit * priceBid + ticketCost) and has some extra amount. Gas = 12500
-const totalCost = 21000; // Arbitration cost + surplus
-
-const L2_GAS_LIMIT = 500;
-const gasPriceBid = 5;
-const l2GasPrice = 12500; // See above how this value calculated
-
-const ticketSubmissionCost = 10000;
 
 const appealTimeOut = 180;
 const winnerMultiplier = 3000;
@@ -37,9 +29,6 @@ let arbitrator;
 let homeProxy;
 let foreignProxy;
 let realitio;
-let mockInbox;
-let mockBridge;
-let mockOutbox;
 
 let governor;
 let requester;
@@ -51,9 +40,7 @@ let other;
 describe("Cross-chain arbitration with appeals", () => {
   beforeEach("initialize the contract", async function () {
     [governor, requester, crowdfunder1, crowdfunder2, answerer, other] = await ethers.getSigners();
-    ({ arbitrator, realitio, foreignProxy, homeProxy, mockInbox, mockBridge, mockOutbox } = await deployContracts(
-      governor
-    ));
+    ({ arbitrator, realitio, foreignProxy, homeProxy } = await deployContracts(governor));
 
     // Create disputes so the index in tests will not be a default value.
     await arbitrator.connect(other).createDispute(42, arbitratorExtraData, { value: arbitrationCost });
@@ -67,146 +54,62 @@ describe("Cross-chain arbitration with appeals", () => {
   it("Should correctly set the initial values", async () => {
     expect(await foreignProxy.arbitrator()).to.equal(arbitrator.target);
     expect(await foreignProxy.arbitratorExtraData()).to.equal(arbitratorExtraData);
-    expect(await foreignProxy.inbox()).to.equal(mockInbox.target);
-    expect(await foreignProxy.l2GasLimit()).to.equal(500);
-    expect(await foreignProxy.gasPriceBid()).to.equal(5);
-    expect(await foreignProxy.surplusAmount()).to.equal(20000);
-    expect(await foreignProxy.homeProxy()).to.equal(homeProxy.target);
-    expect(await foreignProxy.wNative()).to.equal(other);
-
+    expect(await foreignProxy.fxChildTunnel()).to.equal(homeProxy.target);
     expect(await homeProxy.metadata()).to.equal(metadata);
     expect(await homeProxy.foreignChainId()).to.equal(zeroPadValue(toBeHex(5), 32));
     expect(await homeProxy.foreignProxy()).to.equal(foreignProxy.target);
-    expect(await homeProxy.mockInbox()).to.equal(mockInbox.target);
-    expect(await homeProxy.mockBridge()).to.equal(mockBridge.target);
-    expect(await homeProxy.realitio()).to.equal(realitio.target);
+    expect(await homeProxy.fxRootTunnel()).to.equal(foreignProxy.target);
 
     // 0 - winner, 1 - loser, 2 - loserAppealPeriod.
     const multipliers = await foreignProxy.getMultipliers();
     expect(multipliers[0]).to.equal(3000);
     expect(multipliers[1]).to.equal(7000);
     expect(multipliers[2]).to.equal(5000);
-
-    expect(await mockOutbox.sender()).to.equal(homeProxy.target);
-    expect(await mockBridge.outbox()).to.equal(mockOutbox.target);
-
-    expect(await mockInbox.bridge()).to.equal(mockBridge.target);
-    expect(await mockInbox.submissionFee()).to.equal(10000);
   });
 
   it("Should set correct values when requesting arbitration and fire the event", async () => {
     await expect(
-      foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost })
+      foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost - 1 })
     ).to.be.revertedWith("Deposit value too low");
 
-    const requesterAddress = await requester.getAddress();
-    await expect(foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost }))
-      .to.emit(mockInbox, "TicketSent")
-      .withArgs(
-        homeProxy.target,
-        0,
-        ticketSubmissionCost,
-        requesterAddress,
-        requesterAddress,
-        L2_GAS_LIMIT,
-        gasPriceBid
-      )
-      .to.emit(foreignProxy, "RetryableTicketCreated")
-      .withArgs(0) // TicketID
-      .to.emit(foreignProxy, "ArbitrationRequested")
-      .withArgs(questionID, requesterAddress, maxPrevious);
-
-    const arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
-    expect(arbitration[0]).to.equal(1, "Incorrect status of the arbitration after creating a request");
-    expect(arbitration[1]).to.equal(8500, "Deposit value stored incorrectly"); // Surplus (20000) + ArbCost (1000) - ArbitrumFee (10000 + 5*500).
-
-    // Inbox acts as foreignProxyAlias here so it passes onlyForeignProxyAlias require
-    await expect(mockInbox.connect(other).redeemTicket(0))
+    await expect(
+      foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost })
+    )
       .to.emit(realitio, "MockNotifyOfArbitrationRequest")
-      .withArgs(questionID, requesterAddress)
+      .withArgs(questionID, await requester.getAddress())
       .to.emit(homeProxy, "RequestNotified")
-      .withArgs(questionID, requesterAddress, maxPrevious);
+      .withArgs(questionID, await requester.getAddress(), maxPrevious)
+      .to.emit(foreignProxy, "ArbitrationRequested")
+      .withArgs(questionID, await requester.getAddress(), maxPrevious);
 
-    const request = await homeProxy.requests(questionID, requesterAddress);
-    expect(request[0]).to.equal(2, "Incorrect status of the request in HomeProxy");
-    expect(request[1]).to.equal(ZERO_HASH, "Answer should be empty");
-
-    expect(await homeProxy.questionIDToRequester(questionID)).to.equal(
-      requesterAddress,
-      "Incorrect requester stored in home proxy"
-    );
-  });
-
-  it("Should not allow to request arbitration 2nd time", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
+    const arbitration = await foreignProxy.arbitrationRequests(0, await requester.getAddress());
+    expect(arbitration[0]).to.equal(1, "Incorrect status of the arbitration after creating a request");
+    expect(arbitration[1]).to.equal(1000, "Deposit value stored incorrectly");
 
     await expect(
-      foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost })
+      foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost })
     ).to.be.revertedWith("Arbitration already requested");
   });
 
-  it("Should have correct balance after paying arbitration cost and arbitrum fee", async () => {
+  it("Should set correct values when acknowledging arbitration and create a dispute", async () => {
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: oneETH }); // Deliberately overpay
     const oldBalance = await getBalance(requester);
 
-    const tx = await foreignProxy
-      .connect(requester)
-      .requestArbitration(questionID, maxPrevious, { gasPrice: gasPrice, value: totalCost });
-    const txFee = (await tx.wait()).gasUsed * gasPrice;
-
-    const newBalance = await getBalance(requester);
-    expect(newBalance).to.equal(
-      oldBalance - toBigInt(1000) - txFee - toBigInt(20000), // Subtract tx fee, arbitration cost and surplus. The leftover surplus will be reimbursed later
-      "Requester was not reimbursed correctly"
-    );
-  });
-
-  it("Check home proxy permissions", async () => {
     await expect(
-      homeProxy.receiveArbitrationRequest(questionID, await requester.getAddress(), maxPrevious)
-    ).to.be.revertedWith("Can only be called by foreign proxy");
+      foreignProxy.connect(other).receiveArbitrationAcknowledgement(questionID, await requester.getAddress())
+    ).to.be.revertedWith("Can only be called via bridge");
 
-    await expect(homeProxy.receiveArbitrationFailure(questionID, await requester.getAddress())).to.be.revertedWith(
-      "Can only be called by foreign proxy"
-    );
-
-    await expect(homeProxy.receiveArbitrationAnswer(questionID, answer)).to.be.revertedWith(
-      "Can only be called by foreign proxy"
-    );
-  });
-
-  it("Check foreign proxy permissions", async () => {
-    await expect(
-      foreignProxy.receiveArbitrationAcknowledgement(questionID, await requester.getAddress())
-    ).to.be.revertedWith("NOT_BRIDGE");
-
-    await expect(
-      foreignProxy.receiveArbitrationCancelation(questionID, await requester.getAddress())
-    ).to.be.revertedWith("NOT_BRIDGE");
-  });
-
-  it("Should set correct values when acknowledging arbitration and create a dispute", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
-
-    const badMessage = "0xfa";
-    await expect(mockBridge.connect(other).sendAsBridge(homeProxy.target, badMessage)).to.be.revertedWith(
-      "Failed TxToL1"
-    );
-
-    await expect(homeProxy.connect(other).handleNotifiedRequest(questionID, await requester.getAddress()))
-      .to.emit(homeProxy, "RequestAcknowledged")
-      .withArgs(questionID, await requester.getAddress())
+    await expect(homeProxy.handleNotifiedRequest(questionID, await requester.getAddress()))
       .to.emit(arbitrator, "DisputeCreation")
       .withArgs(2, foreignProxy.target)
       .to.emit(foreignProxy, "ArbitrationCreated")
       .withArgs(questionID, await requester.getAddress(), 2)
       .to.emit(foreignProxy, "Dispute")
-      .withArgs(arbitrator.target, 2, 0, 0);
+      .withArgs(arbitrator.target, 2, 0, 0) // Arbitrator, DisputeID, MetaevidenceID, ArbitrationID
+      .to.emit(homeProxy, "RequestAcknowledged")
+      .withArgs(questionID, await requester.getAddress());
 
-    const request = await homeProxy.requests(questionID, await requester.getAddress());
-    expect(request[0]).to.equal(3, "Incorrect status of the request in HomeProxy");
-
-    const arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
+    const arbitration = await foreignProxy.arbitrationRequests(0, await requester.getAddress());
     expect(arbitration[0]).to.equal(2, "Incorrect status of the arbitration after acknowledging arbitration");
     expect(arbitration[1]).to.equal(0, "Deposit value should be empty");
     expect(arbitration[2]).to.equal(2, "Incorrect dispute ID");
@@ -234,248 +137,99 @@ describe("Cross-chain arbitration with appeals", () => {
     expect(dispute[0]).to.equal(foreignProxy.target, "Incorrect arbitrable address");
     expect(dispute[1]).to.equal(MAX_ANSWER, "Incorrect number of choices");
     expect(dispute[2]).to.equal(1000, "Incorrect fees value stored");
-  });
-
-  it("Should not allow to receive the message from incorrect L2 sender", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
-    // Deliberately change outbox sender to see if the check in onlyL2Bridge modifier works.
-    // Note that it want obtain the exact message during revert because of low level call
-    await mockOutbox.setSender(await other.getAddress());
-
-    await expect(
-      homeProxy.connect(other).handleNotifiedRequest(questionID, await requester.getAddress())
-    ).to.be.revertedWith("Failed TxToL1");
-  });
-
-  it("Should not be able to proccess the message twice", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
-
-    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-
-    await expect(
-      foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost })
-    ).to.be.revertedWith("Dispute already created");
-  });
-
-  it("Should not allow to handle request before ticket is redeemed", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-
-    await expect(homeProxy.handleNotifiedRequest(questionID, await requester.getAddress())).to.be.revertedWith(
-      "Invalid request status"
-    );
-  });
-
-  it("Should reimburse after dispute creation in case of overpay", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: oneETH }); // Deliberately overpay
-    await mockInbox.connect(other).redeemTicket(0);
-
-    const arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
-    const requesterDeposit = arbitration[1];
-    expect(requesterDeposit).to.equal(oneETH - toBigInt(12500), "Incorrect deposit value"); //Ticket submission fee + gas limit * price bid = 10000 * 2500
-
-    const oldBalance = await getBalance(requester);
-    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     const newBalance = await getBalance(requester);
     expect(newBalance).to.equal(
-      oldBalance + requesterDeposit - toBigInt(arbitrationCost),
+      oldBalance + oneETH - toBigInt(arbitrationCost),
       "Requester was not reimbursed correctly"
     );
   });
 
   it("Should cancel arbitration correctly", async () => {
-    await expect(homeProxy.handleRejectedRequest(questionID, await requester.getAddress())).to.be.revertedWith(
-      "Invalid request status"
-    );
-
     const badMaxPrevious = 11;
-    await foreignProxy.connect(requester).requestArbitration(questionID, badMaxPrevious, { value: totalCost });
-
-    await expect(mockInbox.connect(other).redeemTicket(0))
+    await expect(foreignProxy.connect(requester).requestArbitration(questionID, badMaxPrevious, { value: 5555 }))
       .to.emit(homeProxy, "RequestRejected")
-      .withArgs(questionID, await requester.getAddress(), badMaxPrevious, "Bond has changed");
-
-    let request = await homeProxy.requests(questionID, await requester.getAddress());
-    expect(request[0]).to.equal(1, "Incorrect status of the request in HomeProxy after rejection");
-    expect(await homeProxy.questionIDToRequester(questionID)).to.equal(
-      ZeroAddress,
-      "Requester address should be empty after rejection"
-    );
+      .withArgs(questionID, await requester.getAddress(), badMaxPrevious, "Bond has changed")
+      .to.emit(foreignProxy, "ArbitrationRequested")
+      .withArgs(questionID, await requester.getAddress(), badMaxPrevious);
 
     const oldBalance = await getBalance(requester);
+    await expect(
+      foreignProxy.connect(other).receiveArbitrationCancelation(questionID, await requester.getAddress())
+    ).to.be.revertedWith("Can only be called via bridge");
+
     await expect(homeProxy.handleRejectedRequest(questionID, await requester.getAddress()))
-      .to.emit(homeProxy, "RequestCanceled")
-      .withArgs(questionID, await requester.getAddress())
       .to.emit(foreignProxy, "ArbitrationCanceled")
+      .withArgs(questionID, await requester.getAddress())
+      .to.emit(homeProxy, "RequestCanceled")
       .withArgs(questionID, await requester.getAddress());
 
-    request = await homeProxy.requests(questionID, await requester.getAddress());
-    expect(request[0]).to.equal(0, "Status should be nullified in home proxy");
-
     const newBalance = await getBalance(requester);
-    expect(newBalance).to.equal(
-      oldBalance + toBigInt(arbitrationCost + surplusAmount - 12500), // 12500 is Arbitrum fee
-      "Requester was not reimbursed correctly"
-    );
+    expect(newBalance).to.equal(oldBalance + toBigInt(5555), "Requester was not reimbursed correctly");
 
-    const arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
+    const arbitration = await foreignProxy.arbitrationRequests(0, await requester.getAddress());
     expect(arbitration[0]).to.equal(0, "Status should be empty");
     expect(arbitration[1]).to.equal(0, "Deposit should be empty");
-    expect(arbitration[2]).to.equal(0, "Dispute id should be empty");
   });
 
   it("Should correctly handle failed dispute creation", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
 
-    await expect(
-      foreignProxy
-        .connect(other)
-        .handleFailedDisputeCreation(questionID, await requester.getAddress(), { value: totalCost })
-    ).to.be.revertedWith("Invalid arbitration status");
+    const oldBalance = await getBalance(requester);
+    await expect(foreignProxy.handleFailedDisputeCreation(questionID, await requester.getAddress())).to.be.revertedWith(
+      "Invalid arbitration status"
+    );
 
-    await arbitrator.setArbitrationPrice(8501); // Increase the cost so creation fails. 21000 was originally sent, 12500 taken as fee, so increased cost should be higher than 8500
+    await arbitrator.setArbitrationPrice(2000); // Increase the cost so creation fails.
 
     await expect(homeProxy.handleNotifiedRequest(questionID, await requester.getAddress()))
       .to.emit(foreignProxy, "ArbitrationFailed")
+      .withArgs(questionID, await requester.getAddress())
+      .to.emit(homeProxy, "RequestAcknowledged")
       .withArgs(questionID, await requester.getAddress());
 
-    let arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
-    expect(arbitration[0]).to.equal(5, "Status should be Failed");
+    let arbitration = await foreignProxy.arbitrationRequests(0, await requester.getAddress());
+    expect(arbitration[0]).to.equal(4, "Status should be Failed");
 
-    const oldBalance = await getBalance(requester);
-
-    await expect(
-      foreignProxy
-        .connect(other)
-        .handleFailedDisputeCreation(questionID, await requester.getAddress(), { value: l2GasPrice })
-    )
-      .to.emit(foreignProxy, "RetryableTicketCreated")
-      .withArgs(1)
-      .to.emit(foreignProxy, "ArbitrationCanceled")
-      .withArgs(questionID, await requester.getAddress())
-      .to.emit(mockInbox, "TicketSent");
-
-    await expect(mockInbox.connect(other).redeemTicket(1))
+    await expect(foreignProxy.handleFailedDisputeCreation(questionID, await requester.getAddress()))
       .to.emit(realitio, "MockCancelArbitrationRequest")
       .withArgs(questionID)
       .to.emit(homeProxy, "ArbitrationFailed")
+      .withArgs(questionID, await requester.getAddress())
+      .to.emit(foreignProxy, "ArbitrationCanceled")
       .withArgs(questionID, await requester.getAddress());
 
     const newBalance = await getBalance(requester);
-    expect(newBalance).to.equal(
-      oldBalance + toBigInt(arbitrationCost + surplusAmount - l2GasPrice),
-      "Requester was not reimbursed correctly"
-    ); // 8500
+    expect(newBalance).to.equal(oldBalance + toBigInt(arbitrationCost), "Requester was not reimbursed correctly");
 
     arbitration = await foreignProxy.arbitrationRequests(0, await requester.getAddress());
-    expect(arbitration[0]).to.equal(5, "Status should be Failed");
+    expect(arbitration[0]).to.equal(0, "Status should be empty");
     expect(arbitration[1]).to.equal(0, "Deposit should be empty");
-    expect(arbitration[2]).to.equal(0, "Dispute id should be empty");
-
-    const request = await homeProxy.requests(questionID, await requester.getAddress());
-    expect(request[0]).to.equal(0, "Status should be nullified in home proxy");
-  });
-
-  it("Should correctly reimburse the overpay when handling failed dispute", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
-    await arbitrator.setArbitrationPrice(8501); // Increase the cost so creation fails.
-    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-
-    await expect(
-      foreignProxy
-        .connect(other)
-        .handleFailedDisputeCreation(questionID, await requester.getAddress(), { value: l2GasPrice - 1 })
-    ).to.be.revertedWith("Should cover arbitrum fee");
-
-    const oldBalance = await getBalance(other);
-    const tx = await foreignProxy
-      .connect(other)
-      .handleFailedDisputeCreation(questionID, await requester.getAddress(), { gasPrice: gasPrice, value: oneETH });
-    const txFee = (await tx.wait()).gasUsed * gasPrice;
-
-    const newBalance = await getBalance(other);
-    expect(newBalance).to.equal(
-      oldBalance - toBigInt(l2GasPrice) - txFee, // Take only gas price for L2 and txfee
-      "Caller was not reimbursed correctly"
-    );
   });
 
   it("Should handle the ruling correctly", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
 
     await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
+    await expect(foreignProxy.rule(2, 8)).to.be.revertedWith("Only arbitrator allowed");
 
     const arbAnswer = zeroPadValue(toBeHex(7), 32);
 
-    await expect(
-      foreignProxy.connect(other).relayRule(questionID, await requester.getAddress(), { value: l2GasPrice })
-    ).to.be.revertedWith("Dispute not resolved");
+    await expect(arbitrator.giveRuling(2, 8))
+      .to.emit(homeProxy, "ArbitratorAnswered")
+      .withArgs(questionID, arbAnswer)
+      .to.emit(foreignProxy, "Ruling")
+      .withArgs(arbitrator.target, 2, 8);
 
-    await expect(arbitrator.giveRuling(2, 8)).to.emit(foreignProxy, "Ruling").withArgs(arbitrator.target, 2, 8);
-
-    let arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
+    const arbitration = await foreignProxy.arbitrationRequests(0, await requester.getAddress());
     expect(arbitration[0]).to.equal(3, "Status should be Ruled");
     expect(arbitration[3]).to.equal(8, "Stored answer is incorrect");
-
-    await expect(homeProxy.reportArbitrationAnswer(questionID, ZERO_HASH, ZERO_HASH, ZeroAddress)).to.be.revertedWith(
-      "Arbitrator has not ruled yet"
-    );
-
-    await expect(foreignProxy.connect(other).relayRule(questionID, await requester.getAddress(), { value: l2GasPrice }))
-      .to.emit(foreignProxy, "RetryableTicketCreated")
-      .withArgs(1)
-      .to.emit(mockInbox, "TicketSent")
-      .to.emit(foreignProxy, "RulingRelayed")
-      .withArgs(questionID, arbAnswer);
-
-    await expect(mockInbox.connect(other).redeemTicket(1))
-      .to.emit(homeProxy, "ArbitratorAnswered")
-      .withArgs(questionID, arbAnswer);
-
-    arbitration = await foreignProxy.arbitrationRequests(arbitrationID, await requester.getAddress());
-    expect(arbitration[0]).to.equal(4, "Status should be Relayed");
-
-    let request = await homeProxy.requests(questionID, await requester.getAddress());
-    expect(request[0]).to.equal(4, "Status should be Ruled");
-    expect(request[1]).to.equal(arbAnswer, "Incorrect answer stored");
 
     await expect(homeProxy.reportArbitrationAnswer(questionID, ZERO_HASH, ZERO_HASH, ZeroAddress))
       .to.emit(realitio, "MockFinalize")
       .withArgs(questionID, arbAnswer)
       .to.emit(homeProxy, "ArbitrationFinished")
       .withArgs(questionID);
-
-    request = await homeProxy.requests(questionID, await requester.getAddress());
-    expect(request[0]).to.equal(5, "Status should be Finished");
-  });
-
-  it("Should correctly reimburse the overpay when ruling is relayed", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
-
-    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-    await arbitrator.giveRuling(2, 8);
-
-    await expect(
-      foreignProxy.connect(other).relayRule(questionID, await requester.getAddress(), { value: l2GasPrice - 1 })
-    ).to.be.revertedWith("Should cover arbitrum fee");
-
-    const oldBalance = await getBalance(other);
-    const tx = await foreignProxy
-      .connect(other)
-      .relayRule(questionID, await requester.getAddress(), { gasPrice: gasPrice, value: oneETH });
-    const txFee = (await tx.wait()).gasUsed * gasPrice;
-
-    const newBalance = await getBalance(other);
-    expect(newBalance).to.equal(
-      oldBalance - toBigInt(l2GasPrice) - txFee, // Take only gas price for L2 and txfee
-      "Caller was not reimbursed correctly"
-    );
   });
 
   it("Should correctly fund an appeal and fire the events", async () => {
@@ -486,15 +240,12 @@ describe("Cross-chain arbitration with appeals", () => {
     let tx;
     let txFee;
     let roundInfo;
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
 
     await expect(foreignProxy.connect(crowdfunder1).fundAppeal(arbitrationID, 11, { value: 1000 })).to.be.revertedWith(
       "No dispute to appeal."
     );
-
     await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
-
     // Check that can't fund the dispute that is not appealable.
     await expect(foreignProxy.connect(crowdfunder1).fundAppeal(arbitrationID, 11, { value: 1000 })).to.be.revertedWith(
       "Appeal period is over."
@@ -573,8 +324,8 @@ describe("Cross-chain arbitration with appeals", () => {
 
   it("Should correctly create and fund subsequent appeal rounds", async () => {
     let roundInfo;
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
+
     await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 21, appealCost, appealTimeOut);
@@ -617,8 +368,8 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should not fund the appeal after the timeout", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
+
     await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 21, appealCost, appealTimeOut);
@@ -649,9 +400,9 @@ describe("Cross-chain arbitration with appeals", () => {
     const crowdfunder1Address = await crowdfunder1.getAddress();
     const crowdfunder2Address = await crowdfunder2.getAddress();
 
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
-    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
+
+    await homeProxy.handleNotifiedRequest(questionID, requesterAddress);
 
     await arbitrator.giveAppealableRuling(2, 5, appealCost, appealTimeOut);
 
@@ -790,9 +541,12 @@ describe("Cross-chain arbitration with appeals", () => {
     await foreignProxy.withdrawFeesAndRewards(arbitrationID, requesterAddress, 2, 41);
     await foreignProxy.withdrawFeesAndRewards(arbitrationID, crowdfunder2Address, 2, 5);
 
-    newBalance =await getBalance(requester);
+    newBalance = await getBalance(requester);
     newBalance2 = await getBalance(crowdfunder2);
-    expect(newBalance).to.equal(oldBalance + toBigInt(8499), "The balance of the requester is incorrect (withdraw 2 round)");
+    expect(newBalance).to.equal(
+      oldBalance + toBigInt(8499),
+      "The balance of the requester is incorrect (withdraw 2 round)"
+    );
     expect(newBalance2).to.equal(
       oldBalance2 + toBigInt(6500),
       "The balance of the crowdfunder2 is incorrect (withdraw 2 round)"
@@ -806,8 +560,8 @@ describe("Cross-chain arbitration with appeals", () => {
     let oldBalance;
     let newBalance;
 
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
+
     await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 20, appealCost, appealTimeOut);
@@ -850,8 +604,8 @@ describe("Cross-chain arbitration with appeals", () => {
     let oldBalance;
     let newBalance;
 
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
+
     await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 3, appealCost, appealTimeOut);
@@ -894,8 +648,8 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should switch the ruling if the loser paid appeal fees while winner did not", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
+
     await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
 
     await arbitrator.giveAppealableRuling(2, 14, appealCost, appealTimeOut);
@@ -910,75 +664,63 @@ describe("Cross-chain arbitration with appeals", () => {
   });
 
   it("Should correctly submit evidence", async () => {
-    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: totalCost });
-    await mockInbox.connect(other).redeemTicket(0);
-    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
 
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
     await expect(foreignProxy.connect(other).submitEvidence(arbitrationID, "text"))
       .to.emit(foreignProxy, "Evidence")
       .withArgs(arbitrator.target, arbitrationID, await other.getAddress(), "text");
+  });
+
+  it("Should forbid requesting arbitration after a dispute has been created for the given question", async () => {
+    await foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost });
+
+    await homeProxy.handleNotifiedRequest(questionID, await requester.getAddress());
+
+    await expect(
+      foreignProxy.connect(requester).requestArbitration(questionID, maxPrevious, { value: arbitrationCost })
+    ).to.be.revertedWith("Dispute already created");
   });
 
   async function deployContracts(signer) {
     const Arbitrator = await ethers.getContractFactory("AutoAppealableArbitrator", signer);
     const arbitrator = await Arbitrator.deploy(String(arbitrationCost));
 
-    const MockOutbox = await ethers.getContractFactory("MockOutbox", signer);
-    const mockOutbox = await MockOutbox.deploy();
-
-    const MockBridge = await ethers.getContractFactory("MockBridge", signer);
-    const mockBridge = await MockBridge.deploy(mockOutbox.target);
-
-    const MockInbox = await ethers.getContractFactory("MockInbox", signer);
-    const mockInbox = await MockInbox.deploy(mockBridge.target, ticketSubmissionCost);
+    const FxRoot = await ethers.getContractFactory("MockFxRoot", signer);
+    const fxRoot = await FxRoot.deploy();
 
     const Realitio = await ethers.getContractFactory("MockRealitio", signer);
     const realitio = await Realitio.deploy();
 
-    const ForeignProxy = await ethers.getContractFactory("RealitioForeignProxyArbitrum", signer);
-    const HomeProxy = await ethers.getContractFactory("MockRealitioHomeProxyArbitrum", signer);
-
-    const address = await signer.getAddress();
-    const nonce = await signer.getNonce();
-
-    const homeProxyAddress = ethers.getCreateAddress({
-      from: address,
-      nonce: nonce + 1, // Add 1 since homeProxy deployment will be after foreignProxy
-    });
+    const ForeignProxy = await ethers.getContractFactory("MockRealitioForeignProxyPolygon", signer);
+    const HomeProxy = await ethers.getContractFactory("MockRealitioHomeProxyPolygon", signer);
 
     const foreignProxy = await ForeignProxy.deploy(
-      other, // Use other address as placeholder for wNative
       arbitrator.target,
       arbitratorExtraData,
       metaEvidence,
       winnerMultiplier,
       loserMultiplier,
       loserAppealPeriodMultiplier,
-      homeProxyAddress,
-      mockInbox.target,
-      surplusAmount,
-      [L2_GAS_LIMIT, gasPriceBid]      
+      ZeroAddress,
+      fxRoot.target
     );
 
     const homeProxy = await HomeProxy.deploy(
       realitio.target,
       metadata,
-      foreignProxy.target,
       foreignChainId,
-      mockInbox.target,
-      mockBridge.target
+      fxRoot.target // Here our mock FxRoot directly calls the FxChildTunnel
     );
 
-    await mockOutbox.setSender(homeProxy.target);
+    await foreignProxy.setFxChildTunnel(homeProxy.target);
+    await homeProxy.setFxRootTunnel(foreignProxy.target);
 
     return {
       arbitrator,
       realitio,
       foreignProxy,
       homeProxy,
-      mockInbox,
-      mockBridge,
-      mockOutbox,
     };
   }
